Process function now working correctly.
Got the process function to work with data being read in from a
datafile, then processed through an ANN, with the output values
being sent into a lms function which produces an lms value.
The lms value produced after each line of data is processed is
added to a running average.

Might need to change the running average to be a mode or something
else in the future.

Currently am running all the datafiles through each ANN, but in
the future if it isnt accurate enough, it could be a good idea
to have different training runs for LOW pressure ANNs, MED
pressure ANNs, and HIGH pressure ANNs. Just thoughts for the
future. So far so good then.

Next step will be to work on the tournament function.