Increased tournament size from 4 to 8, as popsize is 180.
Online article recommends tournament size of 20% of popsize.
So thats size 36, but increased only to 8, as that is at least
better for now, and am still in testing stages.

Also made chances of passing through to next gen without tournament
to 33% chance.

Needed to declare additional individuals in individual aux c file
for tournament, and also refactor tournament code slightly