Added classifier C file, for running classifications
Managed to get a model which looks like it could be effective.
Need a way of now testing such a model, and building a classifier
function will allow to pass in specific data and have the
classifier function return a classification for the data.

Very beginning stages of classifier function, just copied the
process function so far, will refactor it next time.