Refactored process function into normal and oneTime
Copied process function into an additional processBestOneTime()
func due to the large swathes of if else statements within the
process function.

So now the process function is leaner.

Now there is the additional function processBestOneTime which
does similarly to the process function but for a one-off
one-time processing e.g. useful at the end of the algorithm in
order to process data through the best ANN found throughout the
entire run.

Tested and working okay.